% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{single valid number representing a state number in the dataset}

\item{year}{single valid number representing year}
}
\value{
This function returns a base graphics plot outlining the shape of the state with dots overlaying to represent accident occurance (Warning: Function may fail if file does not exist)
}
\description{
Plot Accidents
}
\note{
function is \strong{part of an exercise} in the week2 assignment of a \href{https://coursera.org}{coursera} course titled, Building R Packages

function is for illustrative purposes, \strong{use with care}

function assumes data is from \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{nhtsa.gov}

function assumes data can be found in the working directory

maps package requires a database of maps that can be found in mapdata package
}
\examples{
\dontrun{fars_map_state(1, 2013)}
}

