% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the text the file will read}
}
\value{
This function returns a tbl_df object if file exist. (Warning: Function may fail if file does not exist)
}
\description{
This function makes use of readr's \code{\link[readr]{read_csv}} to read a filename.
}
\note{
function is \strong{part of an exercise} in the week2 assignment of a \href{https://coursera.org}{coursera} course titled, Building R Packages

function is for illustrative purposes, \strong{use with care}

function assumes data is from \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{nhtsa.gov}

function assumes data can be found in the working directory
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read(make_filename(2013))}

}
\seealso{
\code{\link{make_filename}} for making a valid file name, \code{\link{fars_read_years}},
for extension of this functionality
}

