% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Accident}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{valid numeric vector representing year}
}
\value{
This function returns a \code{list} of \code{tbl_df} each with two columns MONTH and year (Warning: Function may fail if file does not exist)
}
\description{
This function loops through a vector of years, read them, and output a \code{data.frame}
containing month and year info
}
\note{
function is \strong{part of an exercise} in the week2 assignment of a \href{https://coursera.org}{coursera} course titled, Building R Packages

function is for illustrative purposes, \strong{use with care}

function assumes data is from \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{nhtsa.gov}

function assumes data can be found in the working directory
}
\examples{
\dontrun{fars_read_years(2013)}
}

