% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count Accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{valid numeric vector representing year}
}
\value{
This function returns a \code{tbl_df} with the first column being MONTH and subsequent onces being the valid years passed into the function (Warning: Function may fail if file does not exist)
}
\description{
This function loops through a vector of years, read them, and summarize the counts (\code{n}) for each year-month combination
}
\note{
function is \strong{part of an exercise} in the week2 assignment of a \href{https://coursera.org}{coursera} course titled, Building R Packages

function is for illustrative purposes, \strong{use with care}

function assumes data is from \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{nhtsa.gov}

function assumes data can be found in the working directory
}
\examples{
\dontrun{fars_summarize_years(2013)}
}

