% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{valid numeric vector representing year}
}
\value{
This function returns a character vector each formatted to "accident_<\code{year}>_.csv.bz2"
}
\description{
This function creates the correct file format to read based on a year parameter
}
\note{
function is \strong{part of an exercise} in the week2 assignment of a \href{https://coursera.org}{coursera} course titled, Building R Packages

function is for illustrative purposes, \strong{use with care}

function assumes data is from \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{nhtsa.gov}

function assumes data can be found in the working directory
}
\examples{
\dontrun{make_filename(2013)}
}

