% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw accidents on a map by year and by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value with the state number to analyse. 
Must be present in FARS data}

\item{year}{A numeric representing the year in 4 digits format YYYY.
Must be present in FARS data}
}
\value{
NULL. The function plots a map with points marking accidents
}
\description{
Draws a map of where the incidents of a given year occurred, 
based on the data of longitude and latitude present in the FARS data. FARS is
US National Highway Traffic Safety Administration's Fatality Analysis 
Reporting System, which is a nationwide census providing the American public yearly data regarding 
fatal injuries suffered in motor vehicle traffic crashes.
}
\note{
the function stops if state.num is not present in FARS data.
}
\examples{
fars_map_state(1,2013)
  
}
