% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with name of the file to read. 
If filename do not exists the function stops execution.}
}
\value{
a data frame tbl (tibble) with data read from file.
}
\description{
This function read the FARS data file using read_csv from readr package using automatic import
for data type. Using read_csv, it reads standard and compressed csv.
}
\note{
The function makes a check if the file exists.
}
\examples{
\dontrun{
    filename <- "accident_2015.csv.bz2" 
    myData <- function(filename)
}
 
}
