% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS files by year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector (atomic or list) of integer(s) representing year in 4 digits format YYYY.}
}
\value{
a list of tibble. If a year is invalid return NULL for that year.
}
\description{
Read one or multiple year from files in the FARS format where FARS is
US National Highway Traffic Safety Administration's Fatality Analysis 
Reporting System, which is a nationwide census providing the American public yearly data regarding 
fatal injuries suffered in motor vehicle traffic crashes.
}
\note{
The function output a warning in case of invalid year. 
      If all years are invalid returns a list of NULL
}
\examples{
fars_read_years(c(2013, 2015))

}
