% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS' years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Numeric vector of FARS´years to read}
}
\value{
A wide format data frame (tibble). One column for year and MONTH. Each value is the count of observations
for the given year-month
}
\description{
This function counts the number of observations by month for the years specified by the \code{years} argument
}
\examples{
  \dontrun{fars_summarize_years(2013:2015)}

}
