% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise number of accident per month and per year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector (atomic or list) of integer(s) representing year in 4 digits format YYYY. 
Years must be presents in FARS data.}
}
\value{
a tibble data frame in the wide format
}
\description{
The function calculate from FARS files the number of accidents per month and per year.
It permits to compare months year by year. FARS is
US National Highway Traffic Safety Administration's Fatality Analysis 
Reporting System, which is a nationwide census providing the American public yearly data regarding 
fatal injuries suffered in motor vehicle traffic crashes.
}
\examples{
fars_summarize_years(c(2013,2014))

}
