% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename for the FARS files}
\usage{
make_filename(year)
}
\description{
This function creates custom filenames based on a year for the FARS files.
}
\details{
@param year Numeric. A number to use for the custom filename

@return A string with the custom filename

@examples
 make_filename(2015)

@export
}
