% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates file name in FARS standard}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric representing the year in 4 digits format YYYY}
}
\value{
A character string with the file name to read in the format accident_YYYY.csv.bz2
}
\description{
Creates file name in the format of FARS files (accident_YYYY.csv.bz2) where YYYY 
is the year in 4 digit format
}
\examples{
make_filename(2013)
   
}
